requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==4.0.0
asyncpg==0.24.0
attrs==21.2.0
charset-normalizer==2.0.7
frozenlist==1.2.0
idna==3.3
iso8601==0.1.16
multidict==5.2.0
pypika-tortoise==0.1.1
pytz==2021.3
tortoise-orm==0.17.8
typing-extensions==3.10.0.2
yarl==1.7.2
